move populateResourceData to generic function
